checking location availability and merging the two sensor record classes
