+++ app/src/main/java/com/dayani/m/roboplatform/RecordSensorsActivity.java
@@ -1,32 +1,49 @@
 package com.dayani.m.roboplatform;
 
-/*
-  Activity to record and store only sensor data:
-       IMU, Magnet, GPS, -> No Camera. -> TODO: Add camera
-  All Sensors are onboard, no external USB device attached to the cellphone.
+/**
+ * Note1: If using multiple threads for handling work
+ *      in each "manager" class in future, use StringBuffer
+ *      instead of StringBuilder (for synchronized operation).
+ * Note2: TODO: better CPU and background managment +
+ *      Better handling critical device states (memory, battery,
+ *      device wake and sleep states, ...).
+ * Note3: TODO: use StringBuffer instead of StringBuilder.
+ *
+ * Note4: When we get here, all permissions are granted and
+ *      USB device is attached and available! So just do
+ *      the checking internally (in each utility class).
  */
 
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
+import android.os.Handler;
+import android.os.HandlerThread;
+import android.os.Looper;
+import android.os.SystemClock;
 import android.util.Log;
 import android.view.View;
+import android.view.WindowManager;
 import android.widget.Button;
+import android.widget.Chronometer;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.content.ContextCompat;
 
+import com.dayani.m.roboplatform.managers.CameraFlyVideo;
 import com.dayani.m.roboplatform.managers.MyLocationManager;
+import com.dayani.m.roboplatform.managers.MySensorManager;
 import com.dayani.m.roboplatform.managers.MyStateManager;
 import com.dayani.m.roboplatform.managers.MyStorageManager;
 import com.dayani.m.roboplatform.managers.MyUSBManager;
 import com.dayani.m.roboplatform.utils.ActivityRequirements;
 import com.dayani.m.roboplatform.utils.AppGlobals;
+import com.dayani.m.roboplatform.utils.AutoFitTextureView;
+import com.dayani.m.roboplatform.utils.SensorsContainer;
 import com.google.android.gms.common.util.ArrayUtils;
 
+import java.io.File;
 import java.util.ArrayList;
 import java.util.Arrays;
 
@@ -35,34 +52,98 @@ public class RecordSensorsActivity extends AppCompatActivity
         implements View.OnClickListener, ActivityRequirements,
         MyUSBManager.OnUsbConnectionListener {
 
-    private static final String TAG = "RecordSensorsActivity";
+    private static final String TAG = "RecordAllActivity";
 
-    public static final String KEY_IS_RECORDING_STATE = AppGlobals.PACKAGE_BASE_NAME
+    private static final String KEY_IS_RECORDING_STATE = AppGlobals.PACKAGE_BASE_NAME
             +".KEY_IS_RECORDING_STATE";
 
-    private static final ArrayList<Requirement> REQUIREMENTS = new ArrayList<>(
-            Arrays.asList(Requirement.PERMISSIONS,
-                    Requirement.ENABLE_LOCATION)
-    );
+    // TODO: Transfer folder management logic to StorageManager Class
+    private static final String VIDEO_FILE_BASE_PATH = "/DCIM/Files";
+    private static final String SENSOR_FILE_BASE_PATH = "/DCIM/Files";
+    private static final String VIDEO_FILE_BASE_NAME = "video_";
+    private static final String SENSOR_FILE_BASE_NAME = "sensor_";
 
+    /**
+     * Button to record video
+     */
     private Button mBtnRecord;
+    private Button mButtonVideo;
+    private Chronometer mChronometer;
+    private AutoFitTextureView mTextureView;
     private TextView reportTxt;
 
+    private MyStorageManager mStore;
+    private File mVideoFile;
+    private File mSensorFile;
+    private String mTimePerfix;
+    private StringBuffer mSensorString;
+
+    private CameraFlyVideo mCam;
     /**
      * Whether the app is recording video now
      */
     private boolean mIsRecording = false;
 
+    private MyUSBManager mUsb;
+
+    private MyLocationManager mLocation;
+
+    private MySensorManager mSensorManager;
+
+    private HandlerThread mSensorThread;
+    private Looper mSensorLooper;
+    private Handler mSensorWriteHandler;
+
+    private Runnable sensorWriteTask = new Runnable() {
+
+        @Override
+        public void run() {
+            Log.d(TAG, "Writing sensor string file in background...");
+
+            //Store sensor values
+            //devise a unified optimum method for setting and
+            //updating compatible video & sensor file names
+            synchronized (this) {
+                mStore.writeBuffered(
+                        new File(MyStorageManager.getNextFilePath(
+                                mSensorFile.getAbsolutePath(),mTimePerfix,"csv")),
+                        mSensorString.toString());
+            }
+        }
+    };
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_record_sensors);
+        setContentView(R.layout.activity_record_all);
 
-        mBtnRecord = findViewById(R.id.record);
-        mBtnRecord.setOnClickListener(this);
+        mButtonVideo = (Button) findViewById(R.id.video);
+        mChronometer = (Chronometer) findViewById(R.id.chronometer);
+        mChronometer.setBase(SystemClock.elapsedRealtime());
+        mButtonVideo.setOnClickListener(this);
         findViewById(R.id.info).setOnClickListener(this);
+        //reportTxt = (TextView) findViewById(R.id.reportText);
+        mTextureView = (AutoFitTextureView) findViewById(R.id.texture);
+
+        mStore = new MyStorageManager(this);
+
+        mVideoFile = mStore.getPublicInternalFile(VIDEO_FILE_BASE_PATH, VIDEO_FILE_BASE_NAME);
+        mSensorFile = mStore.getPublicInternalFile(SENSOR_FILE_BASE_PATH, SENSOR_FILE_BASE_NAME);
+        mTimePerfix = MyStorageManager.getTimePerfix();
+        mSensorString = new StringBuffer();
+
+        mCam = new CameraFlyVideo(this, mTextureView,
+                new File(MyStorageManager.getNextFilePath(
+                        mVideoFile.getAbsolutePath(),mTimePerfix,"mp4")));
+
+        mLocation = new MyLocationManager(this, mSensorString);
+
+        mSensorManager = new MySensorManager(this, mSensorString);
+
+        mUsb = new MyUSBManager(this, this, mSensorString);
+        mUsb.updateDefaultDeviceAvailability();
+
+        startSensorThread();
     }
 
     @Override
@@ -85,7 +166,7 @@ public class RecordSensorsActivity extends AppCompatActivity
         //mUsb.registerUsbPermission();
         //mUsb.tryOpenDefaultDevice();
         //actually, starts preview
-        //mCam.onResume();
+        mCam.onResume();
 
         // Within {@code onPause()}, we remove location updates. Here, we resume receiving
         // location updates if the user has requested them.
@@ -99,7 +180,7 @@ public class RecordSensorsActivity extends AppCompatActivity
     }
 
     @Override
-    protected void onSaveInstanceState(@NonNull Bundle outState) {
+    protected void onSaveInstanceState(Bundle outState) {
         //mLocation.setBundleData(outState);
         // ...
         super.onSaveInstanceState(outState);
@@ -107,14 +188,12 @@ public class RecordSensorsActivity extends AppCompatActivity
 
     @Override
     public void onPause() {
-        //with foreground service we can run this indefinitly so
-        //won't need this
-        //if (mIsRecording) {
-            //stopService();
-        //}
+        if (mIsRecording) {
+            stopRecording();
+        }
         //mUsb.unregisterUsbSensorReciever();
         //mUsb.unregisterUsbPermission();
-        //mCam.clean();
+        mCam.clean();
 
         //mLocation.stopLocationUpdates();
         //mSensorManager.onPause();
@@ -131,12 +210,12 @@ public class RecordSensorsActivity extends AppCompatActivity
     @Override
     protected void onDestroy() {
         //mUsb.unregisterUsbPermission();
-        //mUsb.clean();
+        mUsb.clean();
         //mUsb.close();
         //mUsb = null;
         //cleaning sensor's thread.
         //mSensorManager.clean();
-
+        stopSensorThread();
         super.onDestroy();
     }
 
@@ -144,10 +223,7 @@ public class RecordSensorsActivity extends AppCompatActivity
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
         Log.d(TAG, "onActivityResult");
-        //ToDo: find a way to notify service of this or maybe not
-        //if (mLocation != null) {
-            //mLocation.onActivityResult(requestCode, resultCode, data);
-        //}
+        mLocation.onActivityResult(requestCode, resultCode, data);
     }
 
     @Override
@@ -157,18 +233,18 @@ public class RecordSensorsActivity extends AppCompatActivity
         if (hasFocus) {
             decorView.setSystemUiVisibility(
                     View.SYSTEM_UI_FLAG_LAYOUT_STABLE |
-                            View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY |
-                            View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN |
-                            View.SYSTEM_UI_FLAG_HIDE_NAVIGATION |
-                            View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION |
-                            View.SYSTEM_UI_FLAG_FULLSCREEN
+                    View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY |
+                    View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN |
+                    View.SYSTEM_UI_FLAG_HIDE_NAVIGATION |
+                    View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION |
+                    View.SYSTEM_UI_FLAG_FULLSCREEN
             );
         }
     }
 
      /*
     //Here we don't deal with any permission handling/request &...
-    //If permission are not available, we don't do anything.
+    //If permissions are not available, we don't do anything.
     @Override
     public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
                                            @NonNull int[] grantResults) {
@@ -179,21 +255,60 @@ public class RecordSensorsActivity extends AppCompatActivity
         super.onRequestPermissionsResult(requestCode, permissions, grantResults);
     }*/
 
+    private void startRecording() {
+        //startSensorThread();
+        //Keep the device screen on while recording
+        getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+        //Registers sensor callbacks
+        mSensorManager.start();
+        mUsb.startPeriodicSensorPoll();
+        //while we're not saving files, use our handlerThread for location updates
+        mLocation.startLocationUpdates(mSensorLooper);
+        mCam.startRecordingVideo();
+        mIsRecording = true;
+        MyStateManager.setBoolPref(this, KEY_IS_RECORDING_STATE, mIsRecording);
+        // UI
+        updateButtonState(mIsRecording);
+    }
 
+    private void stopRecording() {
+        //stopSensorThread();
+        //Allow device screen to be turned off
+        getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+        mSensorManager.stop();
+        mUsb.stopPeriodicSensorPoll();
+        mLocation.stopLocationUpdates();
+        mCam.stopRecordingVideo();
+        mIsRecording = false;
+        MyStateManager.setBoolPref(this, KEY_IS_RECORDING_STATE, mIsRecording);
+        //Store sensor values
+        //devise a unified optimum method for setting and
+        //updating compatible video & sensor file names
+        //mStore.writeBuffered(new File(getNextSensorFilePath(mTimePerfix)),
+        //mSensorString.toString());
+        //TODO: or use an asyncTask instead
+        mSensorWriteHandler.post(sensorWriteTask);
+        //preparing the next file
+        mTimePerfix = MyStorageManager.getTimePerfix();
+        mCam.setOutputFile(new File(MyStorageManager.getNextFilePath(
+                mVideoFile.getAbsolutePath(),mTimePerfix,"mp4")));
+        // UI
+        updateButtonState(mIsRecording);
+    }
 
     @Override
     public void onClick(View view) {
         switch (view.getId()) {
-            case R.id.record: {
+            case R.id.video: {
                 if (mIsRecording) {
-                    stopService();
+                    stopRecording();
                 } else {
-                    startService();
+                    startRecording();
                 }
                 break;
             }
             case R.id.info: {
-                Toast.makeText(this,"Record All sensors except camera.",
+                Toast.makeText(this,"Showing sensor information!",
                         Toast.LENGTH_LONG).show();
                 break;
             }
@@ -202,55 +317,61 @@ public class RecordSensorsActivity extends AppCompatActivity
 
     private void updateButtonState(boolean state) {
         if (state) {
-            //mChronometer.setVisibility(View.VISIBLE);
-            //mChronometer.start();
-            mBtnRecord.setText("Stop");
+            mChronometer.setVisibility(View.VISIBLE);
+            mChronometer.start();
+            mButtonVideo.setText("Stop");
             //mButtonVideo.setImageResource(R.drawable.ic_action_pause_over_video);
         } else {
-            //mChronometer.stop();
-            //mChronometer.setVisibility(View.INVISIBLE);
-            mBtnRecord.setText("Start");
+            mChronometer.stop();
+            mChronometer.setVisibility(View.INVISIBLE);
+            mButtonVideo.setText("Start");
             //mButtonVideo.setImageResource(R.drawable.ic_action_play_over_video);
         }
     }
 
-    public void startService() {
-        Intent serviceIntent = new Intent(this, SensorRecordService.class);
-        serviceIntent.putExtra("inputExtra", "Foreground Service Example in Android");
+    /**
+     * Starts a background thread and its {@link Handler}.
+     */
+    private void startSensorThread() {
 
-        ContextCompat.startForegroundService(this, serviceIntent);
-
-        mIsRecording = true;
-        MyStateManager.setBoolPref(this, KEY_IS_RECORDING_STATE, mIsRecording);
-        updateButtonState(mIsRecording);
+        mSensorThread = new HandlerThread(TAG+"SensorBackground");
+        mSensorThread.start();
+        mSensorLooper = mSensorThread.getLooper();
+        mSensorWriteHandler = new Handler(mSensorLooper);
     }
 
-    public void stopService() {
-        Intent serviceIntent = new Intent(this, SensorRecordService.class);
-        stopService(serviceIntent);
+    /**
+     * Stops the background thread and its {@link Handler}.
+     */
+    private void stopSensorThread() {
 
-        mIsRecording = false;
-        MyStateManager.setBoolPref(this, KEY_IS_RECORDING_STATE, mIsRecording);
-        updateButtonState(mIsRecording);
+        mSensorThread.quitSafely();
+        try {
+            mSensorThread.join();
+            mSensorThread = null;
+            mSensorWriteHandler = null;
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
     }
 
     /*============================ Static ActivityRequirements interface =========================*/
-    @Override
-    public boolean usesActivityRequirementsInterface() {
-        return true;
-    }
 
-    public static ArrayList<Requirement> getActivityRequirements() {
-        return REQUIREMENTS;
-    }
+    public static SensorsContainer getSensorRequirements(Context mContext, boolean useUSB) {
 
-    public static String[] getActivityPermissions(Context mContext) {
-        String[] allPermissions = ArrayUtils.concat(MyStorageManager.getPermissions(),
-                MyLocationManager.getPermissions(mContext));
-        Log.i(TAG, Arrays.toString(allPermissions));
-        return allPermissions;
+        SensorsContainer sensors = new SensorsContainer();
+
+        MyStorageManager.getSensorRequirements(mContext, sensors);
+        CameraFlyVideo.getSensorRequirements(mContext, sensors);
+        MyLocationManager.getSensorRequirements(mContext, sensors);
+        MySensorManager.getSensorRequirements(mContext, sensors);
+        if (useUSB) {
+            MyUSBManager.getSensorRequirements(mContext, sensors);
+        }
+
+        //Log.i(TAG, Arrays.toString(sensors.getPermissions().toArray(new String[0])));
+        return sensors;
     }
-    /*============================ Static ActivityRequirements interface =========================*/
 
     public void toastMsgShort(String msg) {
         Toast.makeText(this.getApplicationContext(), msg, Toast.LENGTH_SHORT).show();
