+++ app/src/main/java/com/dayani/m/roboplatform/utils/MySensorInfo.java
@@ -6,6 +6,22 @@ public class MySensorInfo {
 
     }
 
+    public int getId() {
+        return mId;
+    }
+
+    public String getName() {
+        return mName;
+    }
+
+    public String getCalibInfo() {
+        return mCalibInfo;
+    }
+
+    public String getDescInfo() {
+        return mDescInfo;
+    }
+
     private int mId;
     private String mName;
     private String mCalibInfo;
