+++ app/src/main/res/layout/fragment_front_panel.xml
@@ -18,16 +18,16 @@
             android:text="@string/labelStartRecordAll" />
 
         <Button
-            android:id="@+id/startCarManualCtrl"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/labelStartCarManual" />
-
-        <Button
             android:id="@+id/startRecordSensors"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="@string/start_record_sensors" />
+
+        <Button
+            android:id="@+id/startCarManualCtrl"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="@string/labelStartCarManual" />
 
         <Button
             android:id="@+id/startTest"
