+++ app/src/main/res/layout/sensor_group.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+
+    <TextView
+        android:id="@+id/sensor_grp_title"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/sensor_title" />
+
+    <ListView
+        android:id="@+id/sensor_items"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
+</LinearLayout>
